Implement control sequences in door-controller.ini
Sequence allow manipulation of multiple GPIOs when a valid RFID tag is
seen. This could be used to e.g. disable the RFID reader for a few
seconds after each tag, or turn on a warning buzzer when the unlock time
is about to expire (more useful for controlling equipment than a door).

Also, read a host-specific section from the config file to allow for one
file to contain the content for multiple hosts/devices, and/or easier
testing on development machines.