Fix generate-acl.sh when run from install.sh
install.sh runs generate-acls.sh --auth-only, which does something rather
different and has no output directory option. Fix generate-acls.sh not to
capture generate-acls.py's stdout and email it in this case. This is a bit
of a hack since it prevents other options from being used with an output
directory name. Ideally, we'd move all the report generation logic into
the Python script and leave only the mail generation in the shell wrapper,
but I'm too lazy to do that right now.